Extensions + Fix record time
- Add Extensions file and move there some code
- Fix lastRecodeTime label and now it works as it should be
